+++ photons_control/addon.py
@@ -1,7 +1,11 @@
+from photons_app.special import FoundSerials
 from photons_app.actions import an_action
+
+from photons_messages import Services
 
 from delfick_project.addons import addon_hook
 from delfick_project.norms import sb
+import binascii
 
 # Get us our actions
 import photons_control.attributes  # noqa
@@ -51,3 +55,25 @@ async def find_devices(collector, target, reference, artifact, **kwargs):
         _, serials = await reference.find(afr, timeout=20, broadcast=broadcast)
         for serial in serials:
             print(serial)
+
+
+@an_action(needs_target=True)
+async def find_ips(collector, target, reference, **kwargs):
+    """
+    List the ips of the devices that can be found on the network
+
+        lifx lan:find_ips
+
+    You can specify a different broadcast address by saying::
+
+        lifx lan:find_ips 192.168.0.255
+    """
+    broadcast = True
+    if reference not in (None, "", sb.NotSpecified):
+        broadcast = reference
+
+    async with target.session() as afr:
+        found, _ = await FoundSerials().find(afr, timeout=20, broadcast=broadcast)
+        for target, services in found.found.items():
+            if Services.UDP in services:
+                print(f"{binascii.hexlify(target).decode()}: {services[Services.UDP].host}")